+++ src/lib/components/PlayerPresetEditor.svelte
@@ -287,6 +287,7 @@
               {/if}
               {#if draft.systemView !== 'list'}
                 <label>Info panel width (desktop) <span>{draft.inspectorWidth}px</span><input type="range" min="240" max="560" step="10" bind:value={draft.inspectorWidth} /></label>
+                <label>Info panel font size <span>{Math.round(draft.infoFontScale * 100)}%</span><input type="range" min="0.8" max="1.6" step="0.05" bind:value={draft.infoFontScale} /></label>
               {/if}
             </fieldset>
             <fieldset>